microservices-framework/microservices_framework/requirements.txt
appdirs==1.4.4
asgiref==3.3.0
attrs==20.3.0
autopep8==1.5.4
black==20.8b1
botocore==1.19.25
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Django==3.0.2
django-cors-headers==3.6.0
django-filter==2.4.0
django-json-404-middleware==0.0.1
djangorestframework==3.12.2
dnspython==1.16.0
EditorConfig==0.12.2
eventlet==0.30.0
gevent==21.1.0
greenlet==1.0.0
grizzled-python==2.2.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
jmespath==0.10.0
jsbeautifier==1.13.0
Markdown==3.3.3
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pycryptodome==3.9.9
pymongo==3.11.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.0
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.2.4
toml==0.10.2
tornado==6.1
twofish==0.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
uWSGI==2.0.19.1
yapf==0.30.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
